Fixed some gcc warnings beyond -Wall
CVS-ID: phc.c 1.3
CVS-ID: tests.c 1.4
CVS-ID: userom.c 1.4
CVS-ID: yescrypt-common.c 1.2
CVS-ID: yescrypt-opt.c 1.9
CVS-ID: yescrypt-ref.c 1.6

+++ yescrypt-common.c
@@ -428,7 +428,7 @@ uint8_t *yescrypt_r(const yescrypt_shared_t *shared, yescrypt_local_t *local,
 		saltlen = sizeof(saltbin);
 		saltend = decode64(saltbin, &saltlen, saltstr, saltstrlen);
 
-		if (!saltend || saltend - saltstr != saltstrlen)
+		if (!saltend || (size_t)(saltend - saltstr) != saltstrlen)
 			goto fail;
 
 		salt = saltbin;
